+++ src/exactly_lib/instructions/assert_/utils/file_contents/parts/string_matcher_assertion_part.py
@@ -8,7 +8,9 @@ from exactly_lib.symbol.sdv_structure import SymbolReference
 from exactly_lib.test_case.os_services import OsServices
 from exactly_lib.test_case.phases.common import InstructionEnvironmentForPostSdsStep
 from exactly_lib.test_case_utils import pfh_exception
-from exactly_lib.type_system.logic.string_matcher import StringMatcherSdv
+from exactly_lib.type_system.logic.hard_error import HardErrorException
+from exactly_lib.type_system.logic.matching_result import MatchingResult
+from exactly_lib.type_system.logic.string_matcher import StringMatcherSdv, StringMatcher
 from exactly_lib.type_system.logic.string_model import StringModel
 from exactly_lib.util.render import combinators as rend_comb
 
@@ -30,10 +32,21 @@ class StringMatcherAssertionPart(FileContentsAssertionPart):
                custom_environment,
                model: StringModel):
         resolver = resolving_helper_for_instruction_env(os_services, environment)
-        matching_result = resolver.apply(self._string_matcher,
-                                         model)
+        string_matcher = resolver.resolve_matcher(self._string_matcher)
+
+        matching_result = self._apply_matcher(string_matcher, model)
+
         if not matching_result.value:
             raise pfh_exception.PfhFailException(
                 rend_comb.SingletonSequenceR(
                     rendering__node_bool.BoolTraceRenderer(matching_result.trace))
             )
+
+    @staticmethod
+    def _apply_matcher(matcher: StringMatcher,
+                       model: StringModel,
+                       ) -> MatchingResult:
+        try:
+            return matcher.matches_w_trace(model)
+        except HardErrorException as ex:
+            raise pfh_exception.PfhHardErrorException(ex.error)